Add INT casting to convert to number

+++ macros/utils/convert_date_to_part.sql
@@ -1,3 +1,3 @@
 {%- macro convert_date_to_part(datepart, date) -%}
-        EXTRACT({{datepart}} from {{date}}::timestamp)
+        EXTRACT({{datepart}} from {{date}}::timestamp)::int
 {%- endmacro %}
